refactored nav to hook, info, codes, and layout page to components
